REFACTOR abstract classes from 'Support' namespace into 'Sfneal\ViewExport\Support\Adapters' namespace
REFACTOR interfaces from 'Support' namespace into 'Sfneal\ViewExport\Support\Interfaces' namespace.
