+++ src/Excel/Exports/ViewExcelExport.php
@@ -4,7 +4,7 @@ namespace Sfneal\ViewExport\Excel\Exports;
 
 use Illuminate\Contracts\View\View;
 use Maatwebsite\Excel\Concerns\FromView;
-use Sfneal\ViewExport\Support\ExcelExport;
+use Sfneal\ViewExport\Support\Adapters\ExcelExport;
 
 class ViewExcelExport extends ExcelExport implements FromView
 {
