+++ src/Pdf/Utils/PdfExporter.php
@@ -3,9 +3,9 @@
 namespace Sfneal\ViewExport\Pdf\Utils;
 
 use Dompdf\Dompdf;
-use Sfneal\ViewExport\Support\Downloadable;
-use Sfneal\ViewExport\Support\Exporter;
-use Sfneal\ViewExport\Support\Viewable;
+use Sfneal\ViewExport\Support\Adapters\Exporter;
+use Sfneal\ViewExport\Support\Interfaces\Downloadable;
+use Sfneal\ViewExport\Support\Interfaces\Viewable;
 
 class PdfExporter extends Exporter implements Viewable, Downloadable
 {
