--- src/Support/ExcelExport.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Sfneal\ViewExport\Support;
-
-abstract class ExcelExport
-{
-}
